requirements.txt
eve
gunicorn
python-jenkins
